requirements.txt
llvmlite==0.37.0
numba==0.54.1
numpy==1.20.3
opencv-python==4.5.4.60
Pillow==8.4.0
